requirements.txt
cycler==0.10.0
Cython==0.29.21
dask==2.20.0
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.3.0
numpy==1.19.0
pandas==1.0.5
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
scikit-allel==1.3.1
scikit-learn==0.23.1
scipy==1.5.1
six==1.15.0
threadpoolctl==2.1.0
toolz==0.10.0
xgboost==1.1.1
uncertainty-calibration==0.0.7
